+++ internal/testutil/types.go
@@ -31,6 +31,15 @@ func GetClientFromMyContext(ctx MyContext) (graphql.Client, error)     { return 
 const dateFormat = "2006-01-02"
 
 func MarshalDate(t *time.Time) ([]byte, error) {
+	// nil should never happen but we might as well check.  zero-time does
+	// happen because omitempty doesn't consider it zero; we'd prefer to write
+	// null than "0001-01-01".
+	//
+	// (I mean, we're tests.  Who cares!  But we may as well try to match what
+	// prod code would want.)
+	if t == nil || t.IsZero() {
+		return []byte("null"), nil
+	}
 	return []byte(`"` + t.Format(dateFormat) + `"`), nil
 }